Add a low-level HTTP client
This is the main workhorse of the Sensu Go python client. It is
responsible for all communication with the backend while providing
required authentication headers.

Do note that authentication is not part of this base client. This
piece of functionality will be provided by derived clients that will
handle username+password or API key authentication.