Add v2 mixin
Sensu Go python client abstracts away resource structure a bit. We
decided to use the structure of v2 API responses as an internal
representation, which makes transformation from and to the API payload
for the v2 resources trivial.

We will have to do more work for enterprise API endpoints that still
use v1 API version.