Make sure we can create all resources
It turned out that not all resources support creation via the POST
method. This is why we reimplemented the resource creation on top of
the PUT method call.

This did mean that we could inadvertently override an existing
resource on the backend, so we added a check that makes sure no
matching resource exists. This check is not 100% safe because there is
a gap between getting the information and putting the data on the
server, but this is the best we can do.