Add response filtering capability
With the latest addition to the API, we can now filter the list of
resources returned by the list resource client method. Do note that
the filtering is performed by the Sensu Go backend, reducing the
amount of data transferred over the network.

To further reduce the traffic, we made the list result lazy. Resources
are now only downloaded when they are needed. Currently, we download
resources in fixed-size batches of 100 resources per request, but this
can be changed in the future if required.