Validation fix for spec data from API to native
Now validation on spec data happens when we are sending
them to Sensu API. Before, we validated them in both ways
(from API to native and from native to API). Still, it failed
for User resources when we retrieved data as Sensu API
does not return passwords with the user's data.
The solution is only to validate spec data from native to API.